chinook/requirements.txt
matplotlib==3.9.0
seaborn==0.13.2
psycopg2-binary==2.9.9
pandas==2.2.2